cleaned data_classes and removed keys from sample files
