+++ financial_analysis_agent.py
@@ -14,9 +14,9 @@ load_dotenv()
 # Initialize OpenAI API
 openai.api_key = os.getenv("OPENAI_API_KEY")
 catalyst = RagaAICatalyst(
-    access_key="saLy6KmMVlfAzunuQGS9",
-    secret_key="lm39fd4KXffM6gzLjnY9G7QReffhH4RGZPursp3A",
-    base_url="http://52.172.168.127/api"
+    access_key="access_key",
+    secret_key="secret_key",
+    base_url="base_url"
 )
 
 # Start tracing
